+++ crates/bonsaidb-client/src/client/wasm_websocket_worker.rs
@@ -3,7 +3,7 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use bonsaidb_core::{networking::Payload, schema::ApiName};
+use bonsaidb_core::{networking::Payload, api::ApiName};
 use bonsaidb_utils::fast_async_lock;
 use flume::Receiver;
 use url::Url;